requirements.txt
# AWS and Cloud
boto3>=1.26.0
botocore>=1.29.0

# Configuration Management
python-dotenv>=0.21.0
pyyaml>=6.0

# HTTP and Data
requests>=2.28.0
urllib3>=1.26.0

# Development (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=5.0.0
mypy>=0.990

# Utilities
colorama>=0.4.6
tabulate>=0.9.0
